+++ radicale/storage/multifilesystem/__init__.py
@@ -2,7 +2,7 @@
 # Copyright © 2014 Jean-Marc Martins
 # Copyright © 2012-2017 Guillaume Ayoub
 # Copyright © 2017-2021 Unrud <unrud@outlook.com>
-# Copyright © 2024-2025 Peter Bieringer <pb@bieringer.de>
+# Copyright © 2024-2026 Peter Bieringer <pb@bieringer.de>
 #
 # This library is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -214,3 +214,8 @@ class Storage(
                 logger.warning("Storage cache subfolder: %r does not exist, creating now", self._get_collection_cache_folder())
                 self._makedirs_synced(self._get_collection_cache_folder())
             logger.info("Storage cache subfolder permissions: %s", pathutils.path_permissions_as_string(self._get_collection_cache_folder()))
+        if is_collision_free_case_sensitive or self._group_collections_folder is None or len(self._group_collections_folder) == 0:
+            logger.info("Group collections folder: %r", self._group_collections_folder)
+        else:
+            logger.warning("Group collections folder disabled, file system is not case-sensitive: %r", self._group_collections_folder)
+            self._group_collections_folder = ""